requirements.txt
hydra-core==1.3.2
loguru==0.7.2
matplotlib==3.8.0
numpy==1.26.3
omegaconf==2.3.0
onnx==1.16.2
onnx-tf==1.10.0
onnxruntime==1.18.0
onnxruntime-gpu==1.18.0
onnxsim==0.4.33
opencv_python==4.7.0.72
openvino==2024.3.0
openvino_dev==2024.3.0
pandas==2.0.3
Pillow==10.4.0
pillow_heif==0.18.0
pycuda==2024.1
pypdfium2==4.30.0
PyYAML==6.0.2
scikit_learn==1.2.2
tensorrt==8.6.1.post1
torch2trt==0.4.0
timm==1.0.7
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.1
tensorflow==2.9
tensorflow-probability==0.23
tensorflow_addons==0.23
wandb==0.16.6
torchmetrics==1.4.2
